Split angle server request helpers
