+++ frontend/src/server/tools/angle.ts
@@ -3,7 +3,6 @@ import { ApiError, ValidationError } from '@fal-ai/client';
 import { getAngleToolEngine } from '@/config/tools-angle-engines';
 import { isDatabaseConfigured, query, type QueryExecutor, withDbTransaction } from '@/lib/db';
 import { getFalClient } from '@/lib/fal-client';
-import { normalizeMediaUrl } from '@/lib/media';
 import { getResultProviderMode } from '@/lib/result-provider';
 import { ensureBillingSchema } from '@/lib/schema';
 import { computeBillingProductSnapshot } from '@/lib/billing-products';
@@ -18,24 +17,31 @@ import { ensureReusableAsset, upsertLegacyJobOutputs } from '@/server/media-libr
 import {
   applyCinemaSafeParams,
   buildBestAngleVariantParams,
-  mapTiltForEngine,
-  normalizeRotation,
   resolveAngleEngineForParams,
   sanitizeAngleParams,
 } from '@/lib/tools-angle';
 import type {
-  AngleToolEngineDefinition,
   AngleToolOutput,
   AngleToolRequest,
   AngleToolResponse,
   AngleToolEngineId,
 } from '@/types/tools-angle';
 import type { PricingSnapshot } from '@/types/engines';
+import {
+  ANGLE_MULTI_OUTPUT_COUNT,
+  ANGLE_SURFACE,
+  buildAngleFalInput,
+  buildAnglePromptSummary,
+  buildAngleSettingsSnapshot,
+  extractAngleActualCostUsd,
+  extractAngleOutputs,
+  getAngleBillingProductKeyForEngine,
+  parseAngleRequestId,
+  toAngleValidationMessage,
+} from './angle-request-utils';
 
 const TOOL_EVENT_NAME = 'tool_angle_generate';
 const PLACEHOLDER_THUMB = '/assets/frames/thumb-1x1.svg';
-const ANGLE_SURFACE = 'angle' as const;
-const ANGLE_MULTI_OUTPUT_COUNT = 4;
 
 function usdToCredits(value: number | null | undefined): number | null {
   if (typeof value !== 'number' || !Number.isFinite(value) || value <= 0) return null;
@@ -105,44 +111,6 @@ type CreateAngleInitialJobParams = {
   settingsSnapshotJson: string;
   preferredCurrency: Currency | null;
 };
-
-function getAngleBillingProductKeyForEngine(engineId: AngleToolEngineId, generateBestAngles: boolean): string {
-  const family = engineId === 'qwen-multiple-angles' ? 'qwen' : 'flux';
-  return `angle-${family}-${generateBestAngles ? 'multi' : 'single'}`;
-}
-
-function buildAnglePromptSummary(params: { rotation: number; tilt: number; zoom: number }, outputCount: number): string {
-  return `Angle tool · rotation ${Math.round(params.rotation)}° · tilt ${Math.round(params.tilt)}° · zoom ${params.zoom.toFixed(1)} · ${outputCount} output${outputCount > 1 ? 's' : ''}`;
-}
-
-function buildAngleSettingsSnapshot(args: {
-  engine: AngleToolEngineDefinition;
-  imageUrl: string;
-  requested: AngleToolRequest['params'];
-  applied: AngleToolRequest['params'] & { safeMode: boolean; safeApplied: boolean };
-  generateBestAngles: boolean;
-  outputCount: number;
-  billingProductKey: string;
-}): Record<string, unknown> {
-  return {
-    schemaVersion: 1,
-    surface: ANGLE_SURFACE,
-    billingProductKey: args.billingProductKey,
-    engineId: args.engine.id,
-    engineLabel: args.engine.label,
-    inputMode: 'i2i',
-    prompt: buildAnglePromptSummary(args.applied, args.outputCount),
-    source: {
-      imageUrl: args.imageUrl,
-    },
-    controls: {
-      requested: args.requested,
-      applied: args.applied,
-      generateBestAngles: args.generateBestAngles,
-      outputCount: args.outputCount,
-    },
-  };
-}
 
 async function recordAngleRefundReceipt(receipt: PendingAngleReceipt, label: string, priceOnly: boolean) {
   try {
@@ -310,186 +278,6 @@ async function createAtomicInitialAngleJob(params: CreateAngleInitialJobParams):
   }
 }
 
-function normalizeFalUrl(value: string): string {
-  const trimmed = value.trim();
-  if (!trimmed) return trimmed;
-  if (/^https?:\/\//i.test(trimmed)) {
-    return trimmed;
-  }
-  const normalized = trimmed.replace(/^\.?\/+/, '');
-  return `https://fal.media/files/${normalized}`;
-}
-
-function toAngleOutput(entry: Record<string, unknown>): AngleToolOutput | null {
-  const urlRaw =
-    typeof entry.url === 'string'
-      ? entry.url
-      : typeof entry.image_url === 'string'
-        ? entry.image_url
-        : typeof entry.path === 'string'
-          ? entry.path
-          : null;
-  if (!urlRaw) return null;
-
-  const width = typeof entry.width === 'number' ? entry.width : null;
-  const height = typeof entry.height === 'number' ? entry.height : null;
-  const mimeType =
-    typeof entry.content_type === 'string'
-      ? entry.content_type
-      : typeof entry.mimetype === 'string'
-        ? entry.mimetype
-        : typeof entry.mime === 'string'
-          ? entry.mime
-          : null;
-
-  return {
-    url: normalizeMediaUrl(normalizeFalUrl(urlRaw)) ?? normalizeFalUrl(urlRaw),
-    width,
-    height,
-    mimeType,
-  };
-}
-
-function extractOutputs(payload: unknown): AngleToolOutput[] {
-  const roots: unknown[] = [];
-  if (payload && typeof payload === 'object') {
-    const record = payload as Record<string, unknown>;
-    roots.push(record);
-    if (record.output && typeof record.output === 'object') roots.push(record.output);
-    if (record.response && typeof record.response === 'object') roots.push(record.response);
-    if (record.data && typeof record.data === 'object') roots.push(record.data);
-  }
-
-  for (const root of roots) {
-    if (!root || typeof root !== 'object') continue;
-    const record = root as Record<string, unknown>;
-
-    if (Array.isArray(record.images)) {
-      const mapped = record.images
-        .map((entry) => (entry && typeof entry === 'object' ? toAngleOutput(entry as Record<string, unknown>) : null))
-        .filter((entry): entry is AngleToolOutput => Boolean(entry));
-      if (mapped.length) return mapped;
-    }
-
-    if (record.image && typeof record.image === 'object') {
-      const single = toAngleOutput(record.image as Record<string, unknown>);
-      if (single) return [single];
-    }
-
-    if (typeof record.url === 'string') {
-      const single = toAngleOutput(record);
-      if (single) return [single];
-    }
-  }
-
-  return [];
-}
-
-function parseRequestId(payload: unknown): string | undefined {
-  if (!payload || typeof payload !== 'object') return undefined;
-  const record = payload as Record<string, unknown>;
-  const direct = typeof record.request_id === 'string' ? record.request_id : undefined;
-  if (direct) return direct;
-  if (typeof record.id === 'string') return record.id;
-  if (record.response && typeof record.response === 'object') {
-    const responseRecord = record.response as Record<string, unknown>;
-    if (typeof responseRecord.request_id === 'string') return responseRecord.request_id;
-    if (typeof responseRecord.id === 'string') return responseRecord.id;
-  }
-  if (record.output && typeof record.output === 'object') {
-    const outputRecord = record.output as Record<string, unknown>;
-    if (typeof outputRecord.request_id === 'string') return outputRecord.request_id;
-    if (typeof outputRecord.id === 'string') return outputRecord.id;
-  }
-  return undefined;
-}
-
-function extractActualCostUsd(payload: unknown): number | null {
-  const queue: Array<{ value: unknown; depth: number }> = [{ value: payload, depth: 0 }];
-  const candidates: number[] = [];
-
-  const isCostKey = (rawKey: string): boolean => {
-    const key = rawKey.toLowerCase();
-    if (key.includes('zoom_amount')) return false;
-    return (
-      key === 'cost' ||
-      key === 'price' ||
-      key.includes('cost_usd') ||
-      key.includes('price_usd') ||
-      key.includes('actual_cost') ||
-      key.includes('estimated_cost') ||
-      key.endsWith('_usd') ||
-      key.endsWith('usd')
-    );
-  };
-
-  while (queue.length) {
-    const current = queue.shift();
-    if (!current) continue;
-    const { value, depth } = current;
-    if (depth > 5 || value == null) continue;
-
-    if (typeof value === 'string') {
-      const cleaned = value.trim();
-      const dollarMatch = cleaned.match(/\$\s*([0-9]+(?:\.[0-9]+)?)/);
-      if (dollarMatch) {
-        const parsed = Number(dollarMatch[1]);
-        if (Number.isFinite(parsed) && parsed > 0) candidates.push(parsed);
-      }
-      continue;
-    }
-
-    if (Array.isArray(value)) {
-      value.slice(0, 20).forEach((entry) => queue.push({ value: entry, depth: depth + 1 }));
-      continue;
-    }
-
-    if (typeof value === 'object') {
-      const record = value as Record<string, unknown>;
-      Object.entries(record).forEach(([key, field]) => {
-        const costLike = isCostKey(key);
-
-        if (costLike && typeof field === 'number' && Number.isFinite(field) && field > 0 && field < 10_000) {
-          candidates.push(field);
-          return;
-        }
-
-        if (costLike && typeof field === 'string') {
-          const parsed = Number(field.replace(/[^0-9.]+/g, ''));
-          if (Number.isFinite(parsed) && parsed > 0) {
-            candidates.push(parsed);
-            return;
-          }
-        }
-
-        queue.push({ value: field, depth: depth + 1 });
-      });
-    }
-  }
-
-  if (!candidates.length) return null;
-  const costCandidate = candidates.find((value) => value <= 100) ?? candidates[0];
-  return Number(costCandidate.toFixed(6));
-}
-
-function buildFalInput(
-  engine: AngleToolEngineDefinition,
-  imageUrl: string,
-  params: { rotation: number; tilt: number; zoom: number }
-) {
-  const horizontalAngle = Math.round(normalizeRotation(params.rotation));
-  const verticalAngle = Math.round(mapTiltForEngine(engine.id, params.tilt));
-  const zoomAmount = Number(params.zoom.toFixed(1));
-
-  return {
-    image_urls: [imageUrl],
-    horizontal_angle: horizontalAngle,
-    vertical_angle: verticalAngle,
-    zoom_amount: zoomAmount,
-    num_images: 1,
-  };
-}
-
 async function insertToolEvent(params: {
   jobId: string;
   engineId: AngleToolEngineId;
@@ -632,23 +420,6 @@ async function persistAngleOutputs(params: {
   );
 }
 
-function toValidationMessage(error: ValidationError): string {
-  const messages = error.fieldErrors
-    .map((entry) => {
-      const loc = Array.isArray(entry.loc) ? entry.loc.filter((part) => part !== 'body') : [];
-      const path = loc.length ? loc.join('.') : null;
-      const msg = typeof entry.msg === 'string' ? entry.msg.trim() : '';
-      if (!msg) return null;
-      return path ? `${path}: ${msg}` : msg;
-    })
-    .filter((entry): entry is string => Boolean(entry));
-
-  if (!messages.length) {
-    return error.message || 'Validation failed';
-  }
-  return messages.slice(0, 3).join(' · ');
-}
-
 export async function runAngleTool(input: RunAngleToolInput): Promise<AngleToolResponse> {
   const requested = sanitizeAngleParams(input.params);
   const safeMode = input.safeMode !== false;
@@ -741,8 +512,8 @@ export async function runAngleTool(input: RunAngleToolInput): Promise<AngleToolR
   });
 
   const falInputs = generateBestAngles
-    ? buildBestAngleVariantParams(applied).map((variant) => buildFalInput(engine, input.imageUrl, variant))
-    : [buildFalInput(engine, input.imageUrl, applied)];
+    ? buildBestAngleVariantParams(applied).map((variant) => buildAngleFalInput(engine, input.imageUrl, variant))
+    : [buildAngleFalInput(engine, input.imageUrl, applied)];
   const falClient = getFalClient();
   let providerJobId: string | null = null;
   let lastQueueUpdate: unknown = null;
@@ -768,8 +539,8 @@ export async function runAngleTool(input: RunAngleToolInput): Promise<AngleToolR
         },
       });
       results.push(result);
-      outputs.push(...extractOutputs(result.data));
-      const providerCostUsd = extractActualCostUsd(result.data) ?? extractActualCostUsd(lastQueueUpdate) ?? 0;
+      outputs.push(...extractAngleOutputs(result.data));
+      const providerCostUsd = extractAngleActualCostUsd(result.data) ?? extractAngleActualCostUsd(lastQueueUpdate) ?? 0;
       providerCostUsdTotal += providerCostUsd;
     }
 
@@ -783,7 +554,7 @@ export async function runAngleTool(input: RunAngleToolInput): Promise<AngleToolR
     const latencyMs = Date.now() - startedAt;
     const finalResult = results[results.length - 1] ?? null;
     const providerCostUsd = providerCostUsdTotal > 0 ? Number(providerCostUsdTotal.toFixed(6)) : null;
-    const requestId = providerJobId ?? (finalResult ? parseRequestId(finalResult.data) ?? finalResult.requestId ?? null : null);
+    const requestId = providerJobId ?? (finalResult ? parseAngleRequestId(finalResult.data) ?? finalResult.requestId ?? null : null);
     const persistedOutputs = await persistAngleOutputs({
       outputs,
       userId: input.userId,
@@ -943,7 +714,7 @@ export async function runAngleTool(input: RunAngleToolInput): Promise<AngleToolR
       detail = error.detail;
       code = error.code;
     } else if (error instanceof ValidationError) {
-      message = toValidationMessage(error);
+      message = toAngleValidationMessage(error);
       status = 422;
       detail = error.fieldErrors;
       code = 'validation_error';
